+++ CHANGELOG.md
@@ -57,8 +57,10 @@ this format is adapted from [Keep a Changelog](https://keepachangelog.com/en/1.1
     [#448](https://github.com/marbl/MetagenomeScope/issues/448).
 
 - Simplify and improve component sorting procedure: use total node / edge
-  length as a tiebreaker, followed by lexicographically minimum name, followed
-  by number of positive node or edge names
+  length as a tiebreaker, followed by lexicographically minimum node / edge
+  name (ignoring orientation), followed by number of positive node / edge
+  names, followed by lexicographically minimum node / edge name (with
+  orientation).
   ([#378](https://github.com/marbl/MetagenomeScope/issues/378),
   [#418](https://github.com/marbl/MetagenomeScope/issues/418),
   [#451](https://github.com/marbl/MetagenomeScope/issues/451),
@@ -67,6 +69,9 @@ this format is adapted from [Keep a Changelog](https://keepachangelog.com/en/1.1
   - Among other things, this means that drawings including many single-node
     components should now be sorted nicely from long to short (similarly to how
     Bandage does things).
+
+  - Components should also now be sorted deterministically. Probably that
+    was already the case but now it is at least more provably consistent.
 
 - Adjust component tiling -- mostly by adjusting "breakpoint detection"
   ([#452](https://github.com/marbl/MetagenomeScope/issues/452)).